collapse sort-package-json into other tasks
