+++ lib/tasks/npm-scripts.js
@@ -1,9 +1,11 @@
 /* @flow */
 'use strict';
 
+const omit = require('lodash.omit');
 const path = require('path');
 const updateJsonFile = require('update-json-file');
 
+const { removeScript } = require('../pkg.js');
 const { JSON_OPTIONS, PACKAGE_JSON } = require('../values.js');
 
 const LABEL = 'npm scripts configured';
@@ -23,6 +25,12 @@ function ava(pkg) {
   return scripts;
 }
 
+const fmt = () => ({
+  fmt: 'npm run sort-package-json',
+  pretest: 'npm run fmt',
+  'sort-package-json': 'npx -q sort-package-json',
+});
+
 function fn({ cwd } /* : TaskOptions */) {
   return updateJsonFile(
     path.join(cwd, 'package.json'),
@@ -35,8 +43,15 @@ function fn({ cwd } /* : TaskOptions */) {
 
       pkg.scripts = {
         ...ava(safePkg),
+        ...fmt(),
         ...(pkg.scripts || {}),
       };
+
+      pkg.scripts = omit(pkg.scripts, [
+        // drop fixpack
+        'fixpack',
+      ]);
+      removeScript(pkg, 'posttest', 'npm run fixpack');
 
       return pkg;
     },
